feat: api 수정

+++ server/src/main/java/com/hackathon/server/course/dto/CourseRequest.java
@@ -3,10 +3,10 @@ package com.hackathon.server.course.dto;
 import java.util.List;
 
 public record CourseRequest (
-        Integer startX,
-        Integer startY,
-        Integer endX,
-        Integer endY,
+        Double startX,
+        Double startY,
+        Double endX,
+        Double endY,
         String contents
 ){
 }
